Move ActivityKeyer/ActivityTree to Composer
