--- packages/component/src/hooks/internal/useDispatchTranscriptFocus.js
@@ -1,7 +1,0 @@
-import useWebChatUIContext from './useWebChatUIContext';
-
-export default function useDispatchTranscriptFocus() {
-  const { dispatchTranscriptFocus, numTranscriptFocusObservers } = useWebChatUIContext();
-
-  return numTranscriptFocusObservers ? dispatchTranscriptFocus : undefined;
-}
