+++ packages/component/src/hooks/internal/useDispatchTranscriptFocusByActivityKey.ts
@@ -1,0 +1,5 @@
+import useWebChatUIContext from './useWebChatUIContext';
+
+export default function useDispatchTranscriptFocusByActivityKey(): (activityKey: string | undefined) => void {
+  return useWebChatUIContext().dispatchTranscriptFocusByActivityKey;
+}
